+++ android/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>TKELogGPSApp</name>
+	<name>TKELogGPS</name>
 	<comment>Project android created by Buildship.</comment>
 	<projects>
 	</projects>
