+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.tkeloggpsapp",
+    package = "com.tks.tkeloggps",
 )
 
 android_resource(
     name = "res",
-    package = "com.tkeloggpsapp",
+    package = "com.tks.tkeloggps",
     res = "src/main/res",
 )
