+++ android/app/src/main/java/com/tks/tkeloggps/MainActivity.java
@@ -1,4 +1,4 @@
-package com.tkeloggpsapp;
+package com.tks.tkeloggps;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -13,7 +13,7 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "TKELogGPSApp";
+        return "TKELogGPS";
     }
 
     @Override
