+++ android/app/src/main/java/com/tks/tkeloggps/MainApplication.java
@@ -1,4 +1,4 @@
-package com.tkeloggpsapp;
+package com.tks.tkeloggps;
 
 import android.app.Application;
