+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "TKELogGPSApp",
-  "displayName": "TKELogGPSApp"
+  "name": "TKELogGPS",
+  "displayName": "TKELogGPS"
 }
