+++ ios/TKELogGPS-tvOS/Info.plist
