+++ ios/TKELogGPS/AppDelegate.h
