+++ ios/TKELogGPS/Images.xcassets/AppIcon.appiconset/Contents.json
