+++ ios/TKELogGPSTests/TKELogGPSTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface TKELogGPSAppTests : XCTestCase
+@interface TKELogGPSTests : XCTestCase
 
 @end
 
-@implementation TKELogGPSAppTests
+@implementation TKELogGPSTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
