+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "TKELogGPSApp",
+  "name": "TKELogGPS",
   "version": "0.0.1",
   "private": true,
   "scripts": {
